Fix edge-case with nested HTML
There can be a case where we truncate everything (no text is left)
because of HTML pushing everything past max height, but we don't think
we actually truncated anything.  In this case, if we say we didn't
truncate anything, but we're still past the max height (we must have
tried to truncate everything), we'll remove the node entirely.